Refactored metadata form for re-use in annotation view
